move to types.h

+++ src/dilithium/randombytes.c
@@ -1,6 +1,6 @@
 #include <dilithium/randombytes.h>
 #include <libfam/rng.h>
-void randombytes(uint8_t *out, size_t outlen) {
+void randombytes(u8 *out, u64 outlen) {
 	Rng rng;
 	rng_init(&rng, NULL);
 	rng_gen(&rng, out, outlen);
