+++ src/include/dilithium/randombytes.h
@@ -1,9 +1,8 @@
 #ifndef RANDOMBYTES_H
 #define RANDOMBYTES_H
 
-#include <stddef.h>
-#include <stdint.h>
+#include <libfam/types.h>
 
-void randombytes(uint8_t *out, size_t outlen);
+void randombytes(u8 *out, u64 outlen);
 
 #endif
